1. https://github.com/easelinc/tourist - Simple, flexible tours for your app http://easelinc.github.io/tourist/
Tourist.js is a simple library for creating guided tours through your app. It's better suited to complex, single-page apps than websites. One of our main requirements was the ability to control the interface for each step. For example, a step in the tour might need to open a window or menu to work correctly. Tourist gives you hooks to do this.